spec: timer: Unset clock after test
This patch changes rtems_timer_cancel(), rtems_timer_reset() and
rtems_timer_fire_after() validation tests. It adds a call
to UnsetClock() in teardown to set the state of the system back
to where is was before the test.

In addition, in rtems_timer_fire_after() it fixes a typo in
pre-condition requirements.