spec: Add limits to time performance tests
This patch adds limits to all currently existing time performance
tests. When the measured time during the execution of these tests
is within the range specified by these limits, the test result will
be OK.

The limit data is provided for the following architecture and
environments (the tests can only pass when executed in the same
setting):

* SPARC Instruction Set Simulator:
    + GR721RC SMP
    + GR721RC UNI
    + GR740 SMP
    + GR740 UNI

* Hardware boards:
    + GR721RC SMP
    + GR721RC UNI
    + GR740 SMP
    + GR740 UNI

With the exception of GR721RC SMP, the input data of the hardware
boards contained only measurement data for three tests:

* spec/rtems/part/req/perf-get-buffer.yml
* spec/rtems/part/req/perf-get-no-buffer.yml
* spec/rtems/part/req/perf-return-buffer.yml

The other tests where still under development when the test runs
where made.

The test logs (input data) from the simulator runs where taken
around 2021-11-28.

The test logs from the hardware are completed at 2021-11-21 with
the exception of the GR721RC UNI which completed 2021-11-16 and
GR721RC SMP which completed 2021-12-02.